+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/jenkinsci/analysis-model/compare/analysis-model-2.0.2...master)
+## [Unreleased](https://github.com/jenkinsci/analysis-model/compare/analysis-model-2.1.0...master)
+
+## [2.1.0](https://github.com/jenkinsci/analysis-model/compare/analysis-model-2.0.2...analysis-model-2.1.0)
 
 ### Added
 - [JENKINS-51267](https://issues.jenkins-ci.org/browse/JENKINS-51267),